refactor: split Grafana tools packages and tighten constants hygiene
Turn the Grafana tools monolith into per-tool packages with a shared helpers
module, and move Kafka/RabbitMQ/Clerk env names into config/constants while
replacing leftover numeric HTTP status checks with HTTPStatus.
