+++ config/constants/__init__.py
@@ -60,6 +60,7 @@ from config.constants.buzz import (
     BUZZ_PRIVATE_KEY_ENV,
     BUZZ_RELAY_URL_ENV,
 )
+from config.constants.clerk import CLERK_ISSUER_ENV, CLERK_JWKS_URL_ENV
 from config.constants.coralogix import (
     CORALOGIX_API_KEY_ENV,
     CORALOGIX_APPLICATION_NAME_ENV,
@@ -166,6 +167,13 @@ from config.constants.jenkins import (
     JENKINS_API_TOKEN_ENV,
     JENKINS_BASE_URL_ENV,
     JENKINS_USERNAME_ENV,
+)
+from config.constants.kafka import (
+    KAFKA_BOOTSTRAP_SERVERS_ENV,
+    KAFKA_SASL_MECHANISM_ENV,
+    KAFKA_SASL_PASSWORD_ENV,
+    KAFKA_SASL_USERNAME_ENV,
+    KAFKA_SECURITY_PROTOCOL_ENV,
 )
 from config.constants.kubernetes import (
     KUBECONFIG_CONTENT_ENV,
@@ -292,6 +300,15 @@ from config.constants.product import (
     RELEASE_STAGE_BANNER,
     RELEASES_API_URL_ENV,
     UV_RUN_RECURSION_DEPTH_ENV,
+)
+from config.constants.rabbitmq import (
+    RABBITMQ_HOST_ENV,
+    RABBITMQ_MANAGEMENT_PORT_ENV,
+    RABBITMQ_PASSWORD_ENV,
+    RABBITMQ_SSL_ENV,
+    RABBITMQ_USERNAME_ENV,
+    RABBITMQ_VERIFY_SSL_ENV,
+    RABBITMQ_VHOST_ENV,
 )
 from config.constants.rds import RDS_DB_INSTANCE_IDENTIFIER_ENV, RDS_REGION_ENV
 from config.constants.redis import (
@@ -481,6 +498,8 @@ __all__ = [
     "BUZZ_PATH_ENV",
     "BUZZ_PRIVATE_KEY_ENV",
     "BUZZ_RELAY_URL_ENV",
+    "CLERK_ISSUER_ENV",
+    "CLERK_JWKS_URL_ENV",
     "ATTACHMENT_MAX_FILE_CHARS",
     "ATTACHMENT_MAX_TOTAL_CHARS",
     "CREDITS_DENIED_MESSAGE",
@@ -575,6 +594,11 @@ __all__ = [
     "JENKINS_API_TOKEN_ENV",
     "JENKINS_BASE_URL_ENV",
     "JENKINS_USERNAME_ENV",
+    "KAFKA_BOOTSTRAP_SERVERS_ENV",
+    "KAFKA_SASL_MECHANISM_ENV",
+    "KAFKA_SASL_PASSWORD_ENV",
+    "KAFKA_SASL_USERNAME_ENV",
+    "KAFKA_SECURITY_PROTOCOL_ENV",
     "KUBECONFIG_CONTENT_ENV",
     "KUBECONFIG_CONTEXT_ENV",
     "KUBECONFIG_NAMESPACE_ENV",
@@ -668,6 +692,13 @@ __all__ = [
     "POSTHOG_PERSONAL_API_KEY_ENV",
     "POSTHOG_PROJECT_ID_ENV",
     "POSTHOG_TIMEOUT_SECONDS_ENV",
+    "RABBITMQ_HOST_ENV",
+    "RABBITMQ_MANAGEMENT_PORT_ENV",
+    "RABBITMQ_PASSWORD_ENV",
+    "RABBITMQ_SSL_ENV",
+    "RABBITMQ_USERNAME_ENV",
+    "RABBITMQ_VERIFY_SSL_ENV",
+    "RABBITMQ_VHOST_ENV",
     "RDS_DB_INSTANCE_IDENTIFIER_ENV",
     "RDS_REGION_ENV",
     "REDIS_DATABASE_ENV",
