+++ config/constants/clerk.py
@@ -1,0 +1,11 @@
+"""Clerk authentication environment variable names."""
+
+from __future__ import annotations
+
+CLERK_ISSUER_ENV = "CLERK_ISSUER"
+CLERK_JWKS_URL_ENV = "CLERK_JWKS_URL"
+
+__all__ = [
+    "CLERK_ISSUER_ENV",
+    "CLERK_JWKS_URL_ENV",
+]
