+++ config/constants/kafka.py
@@ -1,0 +1,17 @@
+"""Kafka environment variable names."""
+
+from __future__ import annotations
+
+KAFKA_BOOTSTRAP_SERVERS_ENV = "KAFKA_BOOTSTRAP_SERVERS"
+KAFKA_SECURITY_PROTOCOL_ENV = "KAFKA_SECURITY_PROTOCOL"
+KAFKA_SASL_MECHANISM_ENV = "KAFKA_SASL_MECHANISM"
+KAFKA_SASL_USERNAME_ENV = "KAFKA_SASL_USERNAME"
+KAFKA_SASL_PASSWORD_ENV = "KAFKA_SASL_PASSWORD"
+
+__all__ = [
+    "KAFKA_BOOTSTRAP_SERVERS_ENV",
+    "KAFKA_SASL_MECHANISM_ENV",
+    "KAFKA_SASL_PASSWORD_ENV",
+    "KAFKA_SASL_USERNAME_ENV",
+    "KAFKA_SECURITY_PROTOCOL_ENV",
+]
