+++ config/constants/rabbitmq.py
@@ -1,0 +1,21 @@
+"""RabbitMQ environment variable names."""
+
+from __future__ import annotations
+
+RABBITMQ_HOST_ENV = "RABBITMQ_HOST"
+RABBITMQ_MANAGEMENT_PORT_ENV = "RABBITMQ_MANAGEMENT_PORT"
+RABBITMQ_USERNAME_ENV = "RABBITMQ_USERNAME"
+RABBITMQ_PASSWORD_ENV = "RABBITMQ_PASSWORD"
+RABBITMQ_VHOST_ENV = "RABBITMQ_VHOST"
+RABBITMQ_SSL_ENV = "RABBITMQ_SSL"
+RABBITMQ_VERIFY_SSL_ENV = "RABBITMQ_VERIFY_SSL"
+
+__all__ = [
+    "RABBITMQ_HOST_ENV",
+    "RABBITMQ_MANAGEMENT_PORT_ENV",
+    "RABBITMQ_PASSWORD_ENV",
+    "RABBITMQ_SSL_ENV",
+    "RABBITMQ_USERNAME_ENV",
+    "RABBITMQ_VERIFY_SSL_ENV",
+    "RABBITMQ_VHOST_ENV",
+]
